+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,7 +1,6 @@
 package kitchenpos.product.domain;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 
 @Entity
 public class Product {
@@ -10,17 +9,17 @@ public class Product {
     private Long id;
     @Column(unique = true)
     private String name;
-    private BigDecimal price;
+    private int price;
 
     public Product() {
     }
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, int price) {
         this.name = name;
         this.price = price;
     }
 
-    public Product(Long id, String name, BigDecimal price) {
+    public Product(Long id, String name, int price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -42,11 +41,11 @@ public class Product {
         this.name = name;
     }
 
-    public BigDecimal getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(final BigDecimal price) {
+    public void setPrice(final int price) {
         this.price = price;
     }
 }
