+++ src/test/java/kitchenpos/factory/MenuFixtureFactory.java
@@ -6,11 +6,11 @@ import kitchenpos.menuGroup.domain.MenuGroup;
 import java.math.BigDecimal;
 
 public class MenuFixtureFactory {
-    public static Menu createMenu(Long id, String name, BigDecimal price, MenuGroup menuGroup) {
+    public static Menu createMenu(Long id, String name, int price, MenuGroup menuGroup) {
         return new Menu(id, name, price, menuGroup);
     }
 
-    public static Menu createMenu(String name, BigDecimal price, MenuGroup menuGroup) {
+    public static Menu createMenu(String name, int price, MenuGroup menuGroup) {
         return new Menu(name, price, menuGroup);
     }
 }
