+++ src/test/java/kitchenpos/factory/ProductFixtureFactory.java
@@ -5,11 +5,11 @@ import kitchenpos.product.domain.Product;
 import java.math.BigDecimal;
 
 public class ProductFixtureFactory {
-    public static Product createProduct(Long id, String name, BigDecimal price) {
+    public static Product createProduct(Long id, String name, int price) {
         return new Product(id, name, price);
     }
 
-    public static Product createProduct(String name, BigDecimal price) {
+    public static Product createProduct(String name, int price) {
         return new Product(name, price);
     }
 }
